Wed, Aug 2, 2023, 5:01 PM +03:00

+++ sentimental-readability/readability.py
@@ -30,7 +30,7 @@ grade = (0.0588 * L) - (0.296 * S) - 15.8
 # # print(L)
 # print(S)
 # print(actual_grade)
-actual_grade = round(actual_grade)
+actual_grade = int(round(actual_grade))
 
 if actual_grade > 16:
     print("Grade 16+")
